Add possibility to specify types and typedMatch function on regexp objects.
This is the first step towards achieving https://github.com/strongloop/express/issues/2756.

/:foo{<type>}     - adds a valType=<type> property to the key object
                  - sets the pattern to match floats for 'number' and
                    'float'
                  - sets the pattern to match integers for 'integer'

<type> ::= string | float | number | integer

A new function is added as typedMatch(str) to the regexp objects
returned by pathToRegexp(...). It is a wrapper on top of String.match(regexp),
and it uses the keys property to change the types of the matching groups
accordingly. The changes happen for "float", "number", and "integer".

The pattern is only changed when a different pattern is not defined.
For example, /:foo([123]){integer} will have the pattern [123], but will
also benefit from automatic type changes when using typedMatch.